+++ test/MultRewriteTest.cpp
@@ -11,7 +11,7 @@
 
 /// Check to ensure that the AST testing generator works as expected.
 TEST(MultRewriteTest, astTestingGeneratorTest) { /* NOLINT */
-  const int HIGHEST_CASE_NUM = AstTestingGenerator::getLargestId();
+  const size_t HIGHEST_CASE_NUM = AstTestingGenerator::getLargestId();
   int i = 1;
   for (; i < HIGHEST_CASE_NUM; i++) {
     Ast ast;
